consolidate tools and utils into single folder
