+++ src/tools/auto/logger.js
